fix plugin imports (sort of)
In this jumbled mess of spaghetti code, it is really difficult to work at all, so I am currently refactoring some code and am making changes to stop VSCode from showing a couple hundred errors per file. Most issues are caused by missing typehinting and mis-interpreted function outputs based on these. Whilst trying to fix that, I stumbled across the problem that despite what the docs say, `CipherAPI` is not part of `cipher.plugins` but instead part of `cipher.api`. This commit tries to fix that. Because of the changes I am currently making, I can not fully test this commit, as the version of CipherOS I currently have is not working.