fix "netscan" command
The commadn would always wait 4 seconds and then speed through the list of hosts and report them all as down. This can't be true however, since I am certain that I have an internet connection and I can verify that my network has running hosts with the `nmap` command. I started digging and soon realised that the `ping3` library requires root privileges in order to run. Trying to start CipherOS using `sudo` did not work out however, so I decided to just replace the library completely with the already installed library `subprocess` and the command `ping`, which is built into most operating systems.
This change removes the need for:
 - root privileges
 - a stupid package that does not work
While changing that, I also noticed a variable called `max_workers` (inside the `scannet()` function), which limits the ammount of tests/pings that get executed in parallel. Since my CPU does not have many cores, but my netmask is set to `255.255.0.0` or `16`, I would have to wait an eternity to get the result of `netscan`. Luckily, this limitation is also useless, because the new ping function I coded actually works, meaning that most of the processing time spent by a thread is spent waiting for a ping-response, not for anything meaningful. Because no serious processing is done, the `max_workers` can technically be set to as high of a number as you like. So that's what I did. No more checking cores, because that is not needed at all. No more eternal waiting for the command to run either. I set `max_workers` to `256` for now, so a subnetmask of `255.255.255.0` or `8` finishes the scan in one go. With the current ping-timeout, that takes about 4 seconds.

I have not been able to test the command's functionality on windows or macOS, as I only use linux, but I checked the online documentation and macOS should use the same `ping` command as linux (or one very similar to it) and for windows I use a seperate command-string.
If there are any problems with my implementation on other OSes, I'll have to adjust the commands, but for now this solution is way better than the old one.

If you couldn't notice, I am a bit mad at the code prior to my changes, as most design-choices taken do not make any sense to me what so ever.
I do not like installing and importing packages that don't even do what they are supposed to do.

TL;DR:
I fixed the netscan command. Now it and it's configuration make sense. Only tested on linux.