change `api.starterdir` to `api.configdir`
Storing all the files in the home folder is a bad idea for organization. There is a dedicated folder for storing configs and the likes: the `~/.config/`-folder.
Because the `api.starterdir` is actually a variable, storing the path to the config-dir, I renamed it to `api.configdir`, as this new name is more fitting.
macOS may want to copy the changes I did, but honestly I could not care less about how they want their directories to be sorted.