add plugin "fcommands"
This is a plugin I have been wanting to create. It uses the tool [`fzf`](https://github.com/junegunn/fzf) for various user input related tasks.

Note: I would like to have some way of inserting text into the prompt that is editable by the user, so a command found with `fhist` can be completely edited, but for now I guess I have to keep dreaming, as finding an easy way to get this without more shitty packages/requirements does not seem possible as of now.