make selected command editable
So the note, saying this is not currently possible was a lie. There is a package that does an excellent job at getting user input. And it is already installed too, as it is required in `main.py`. Using `prompt_toolkit.promt()`, I now have the option to have a pre-pasted text, being the selected command.

(I am happy :D)