speed up loading time of plugin "hacknet"
Basically, on my system, the "hacknet" plugin always took multiple seconds, however I didn't want to wait a couple of seconds every time I restart CipherOS. I looked into it and found the issue: when loading, the plugin does some networking stuff to get the ip of the machine and the isp. My internet is a little slow though, so I have to wait every time the plugin loads.
The fix lets the values be unset (technically set to `None`) untill they are needed. As soon as they are needed, they are initialized.
This does not remove the load time, but it moves the load times away from CipherOS's startup and to when the values are actually needed. (Technically, if the plugin is not used, it removes the load time for those values entirely.)