remove need for `sudo` on installation
This botherd me, because I usually don't trust shell-scripts to be run with `sudo`. I looked at why that would be needed and found I could modify the script to not use root privileges at all. This is done by installing the binary to a user-owned folder and adding that to the `$PATH`-variable. To make the change to `$PATH` persistent, the change needs to be added to the auto-execution for the users shell. As of now, only `bash` and `zsh` are supported, but all shells that have a similar behaviour to these and uses a file for auto-execution, it can be added to this install script.