fix dependency-exists-checkk
The old version was checking, if there was a folder with the exact name of the required dependency. This is not sufficient (and my approach technically isn't either), because package-folders can have different names than the package itself. For example with the dependency `PyInputPlus`, the respective folder was named `PyInputPlus-0.2.12`. This does not match, so the API re-downloaded the dependency every time CipherOS loaded.