rename `exit` function
generally, it is a bad idea to have any variable, class, function or similar to overwrite the builtin functions, as one may not expect the change when thinking about using the builtin. that is also the reason i renamed the argument `type` of the argparser to `argtype`, as `type()` is a builtin function.